src/requirements_api.txt
fastapi==0.95.1
fastapi_restful
mangum==0.17.0
psycopg2-binary==2.9.5
pydantic==1.10.7
requests==2.31.0
sqlalchemy==2.0.13
starlette==0.26.1
unidecode==1.3.6
urllib3==1.26.16
